more docs

+++ acore_server_monitoring_measurement/cli/impl.py
@@ -1,6 +1,11 @@
 # -*- coding: utf-8 -*-
 
+"""
+CLI command business logic implementation.
+"""
+
 import pynamodb_mate.api as pm
+from pathlib import Path
 from acore_server_metadata.api import Server
 
 from ..localmetry import (
@@ -9,12 +14,15 @@ from ..localmetry import (
 from ..paths import path_env_name_cache
 
 
-try:
-    env_name = path_env_name_cache.read_text()
-except FileNotFoundError:
-    server = Server.from_ec2_inside()
-    env_name = server.env_name
-    path_env_name_cache.write_text(env_name)
+if Path("/home/ubuntu").exists():
+    try:
+        env_name = path_env_name_cache.read_text()
+    except FileNotFoundError:  # pragma: no cover
+        server = Server.from_ec2_inside()
+        env_name = server.env_name
+        path_env_name_cache.write_text(env_name)
+else:
+    env_name = "sbx"
 
 
 class WorldServerStatusMeasurement(Base):
@@ -25,4 +33,7 @@ class WorldServerStatusMeasurement(Base):
 
 
 def measure_worldserver():
+    """
+    Measure worldserver status once.
+    """
     WorldServerStatusMeasurement.measure_on_worldserver_ec2()
